Revert "Consolidate domain exports into domain indexes"
This reverts commit 3ca8814091cf685f9fecb2043270838780c876a3.
