+++ src/app.js
@@ -1,10 +1,13 @@
-import 'dotenv/config';
 import express from 'express';
 import { StatusCodes } from 'http-status-codes';
 import cookieParser from 'cookie-parser';
-import { sequelize, logger } from './config/index.js';
-import { authHandler, userHandler } from './user/index.js';
-import { routeLogger, handleError } from './middleware/index.js';
+import sequelize from './config/database.js';
+import accounts from './user/auth/auth-handlers.js';
+import users from './user/user-handlers.js';
+import 'dotenv/config';
+import logger from './config/logger.js';
+import routeLogger from './middleware/logger.js';
+import handleError from './middleware/error-handler.js';
 
 process.on('uncaughtException', (err) => {
   handleError(err);
@@ -32,8 +35,8 @@ app.use(express.static('./src/public'));
 app.set('views', './src/views');
 app.set('view engine', 'ejs');
 app.use(routeLogger);
-app.use(authHandler);
-app.use('/user', userHandler);
+app.use(accounts);
+app.use('/user', users);
 app.use((req, res) => {
   res.status(StatusCodes.NOT_FOUND).send('Not Found');
 });
