--- src/middleware/index.js
@@ -1,6 +1,0 @@
-import * as auth from './auth.js';
-import * as errorHandler from './error-handler.js';
-import * as routeLogger from './logger.js';
-import * as validate from './validate.js';
-
-export { auth, errorHandler, routeLogger, validate };
