--- src/user/index.js
@@ -1,6 +1,0 @@
-import * as userHandler from './user-handlers.js';
-import * as userModel from './user-model.js';
-import * as userService from './user-service.js';
-import * as authHandler from './auth/auth-handlers.js';
-
-export { userHandler, userModel, userService, authHandler };
