+++ src/user/user-model.js
@@ -2,7 +2,9 @@ import Joi from 'joi';
 import bcrypt from 'bcrypt';
 import { StatusCodes } from 'http-status-codes';
 import { Model, DataTypes } from 'sequelize';
-import { AppError, sequelize, logger } from '../config/index.js';
+import sequelize from '../config/database.js';
+import logger from '../config/logger.js';
+import AppError from '../config/error.js';
 
 class User extends Model {}
