+++ src/user/user-service.js
@@ -2,7 +2,8 @@ import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import { StatusCodes } from 'http-status-codes';
 import User from './user-model.js';
-import { AppError, logger } from '../config/index.js';
+import logger from '../config/logger.js';
+import AppError from '../config/error.js';
 
 /**
  * Creates a new user
